Override assert_xxx methods to return true
Many tests are using the syntax "expects(xxx).with { block }" and
setting assertions within the block. The behavior of the assertions has
changed over time. Apparently, when the tests were written the
'assert_xxx' methods would return a truthy value but in the current
version of TestUnit they return `nil`. The methods below restore the
original behavior (otherwise we need to change a LOT of tests...)